change TSelectSlide.Lines to an integer. using a byte for this is irrelevant optimization

+++ src/menu/UMenuSelectSlide.pas
@@ -70,7 +70,7 @@ type
       PData:            ^integer;
 
       //For automatically Setting LineCount
-      Lines: byte;
+      Lines: integer;
 
       //Arrows on/off
       showArrows: boolean;      //default is false
